Move to 'util/gitfs'
