+++ lib/platform/gitea/index.spec.ts
@@ -18,8 +18,8 @@ import {
 } from '../../constants/error-messages';
 import { logger as _logger } from '../../logger';
 import { BranchStatus } from '../../types';
+import * as _gitfs from '../../util/gitfs';
 import { setBaseUrl } from '../../util/http/gitea';
-import * as _gitfs from '../git';
 import * as ght from './gitea-helper';
 
 describe('platform/gitea', () => {
@@ -157,13 +157,13 @@ describe('platform/gitea', () => {
     jest.resetModules();
     jest.clearAllMocks();
     jest.mock('./gitea-helper');
-    jest.mock('../git');
+    jest.mock('../../util/gitfs');
     jest.mock('../../logger');
 
     gitea = await import('.');
     helper = (await import('./gitea-helper')) as any;
     logger = (await import('../../logger')).logger as any;
-    gitfs = require('../git');
+    gitfs = require('../../util/gitfs');
     gitfs.isBranchStale.mockResolvedValue(false);
     gitfs.getBranchCommit.mockResolvedValue(mockCommitHash);
