+++ lib/util/gitfs/index.ts
@@ -1,1 +1,2 @@
 export * from './fs';
+export * from './git';
